Move dlopen logic to new LinuxLibraryLoader
